refactor(core): re-home the pytest/testing cluster into core/testing/
Closes CORE-TESTING-REHOME-1: the 11 testing-cluster modules (test_runner,
test_runner_cache, test_runner_outputs, test_impact, test_performance,
pytest_orchestration, pytest_profiles, pytest_marker_exprs,
public_matrix_receipt, project_test_matrix, coverage_policy - ~2.9k lines)
move from the flat infrastructure/core/ top level into an
infrastructure/core/testing/ package with README/AGENTS.

- old infrastructure/core/<module>.py paths remain as explicit re-export
  shims (import parity verified for every externally-imported name; the
  five externally consumed privates are pinned in the moved modules'
  __all__ so the shims re-export them)
- determinism.py (17 general importers) and script_discovery.py (Stage-02
  analysis discovery, not test machinery) deliberately stay at the core
  top level
- core AGENTS.md gains a re-home pointer; testing/ carries README+AGENTS
