+++ infrastructure/core/coverage_policy.py
@@ -1,32 +1,15 @@
-"""Coverage plugin capability probes (Layer 1 — no reporting imports)."""
+"""Backwards-compat shim: the module moved to ``infrastructure.core.testing.coverage_policy``.
 
-from __future__ import annotations
+Old import paths keep resolving (``CORE-TESTING-REHOME-1``); new code imports
+from ``infrastructure.core.testing.coverage_policy`` directly.
+"""
 
-import subprocess
+from infrastructure.core.testing.coverage_policy import (  # noqa: F401
+    PYTEST_HELP_PROBE_TIMEOUT_SECONDS,
+    check_cov_datafile_support,
+)
 
-from infrastructure.core.runtime.environment import get_python_command
-
-PYTEST_HELP_PROBE_TIMEOUT_SECONDS = 30
-
-
-def check_cov_datafile_support() -> bool:
-    """Return True if pytest-cov supports the ``--cov-datafile`` flag."""
-    try:
-        cmd = get_python_command() + ["-m", "pytest", "--help"]
-        # Importing pytest can exceed ten seconds on a loaded macOS worker
-        # while several xdist workers start concurrently. Keep the probe
-        # bounded, but do not turn normal startup contention into a false
-        # "pytest-cov is unsupported" result.
-        result = subprocess.run(
-            cmd,
-            capture_output=True,
-            text=True,
-            timeout=PYTEST_HELP_PROBE_TIMEOUT_SECONDS,
-            check=False,
-        )
-        return "--cov-datafile" in result.stdout
-    except (subprocess.TimeoutExpired, subprocess.CalledProcessError, FileNotFoundError):
-        return False
-
-
-__all__ = ["PYTEST_HELP_PROBE_TIMEOUT_SECONDS", "check_cov_datafile_support"]
+__all__ = [
+    "PYTEST_HELP_PROBE_TIMEOUT_SECONDS",
+    "check_cov_datafile_support",
+]
