+++ infrastructure/core/testing/__init__.py
@@ -1,0 +1,7 @@
+"""Testing-cluster subpackage: pytest execution, profiles, receipts, and probes.
+
+Moved from the flat ``infrastructure/core/`` top level by
+``CORE-TESTING-REHOME-1`` (see ``infrastructure/core/testing/README.md``).
+The old module paths remain as backwards-compat shims that re-export this
+package's public surface; new code imports from here directly.
+"""
